// fiftyone/miris-viewer/js/src/looker-3d/types.ts
export type OverlayLabel = { _id: string; path: string; selected: boolean; color?: string; label?: string; sampleId?: string; _cls: string; instance?: { _cls: "Instance"; _id: string }; } & Record<string, unknown>;
// ...
export type Archetype3d = "point" | "cuboid" | "polyline" | "annotation-plane";